Moved windbarb series to subfolder.
